v2 stream provider
